// front-end/src/common/enums.ts
export const enum DialogIcons {
    Success = "https://inviton-cdn.azureedge.net/assets/img/modal/modal_success_icon.png",
    Info = "https://inviton-cdn.azureedge.net/assets/img/modal/modal_info_icon.png",
    Error = "https://inviton-cdn.azureedge.net/assets/img/modal/modal_err_icon.png",
    Warning = "https://inviton-cdn.azureedge.net/assets/img/modal/modal_warn_icon.png",
    Question = "https://inviton-cdn.azureedge.net/assets/img/modal/modal_question_icon.png",
    Prompt = "https://inviton-cdn.azureedge.net/assets/img/modal/modal_prompt_icon.png",
}